Rename package to fisher & fisher-fresco module
