+++ app/src/main/java/com/github/skydoves/fisherdemo/theme/Color.kt
@@ -14,18 +14,13 @@
  * limitations under the License.
  */
 
-package com.github.skydoves.frescomposabledemo.model
+package com.github.skydoves.fisherdemo.theme
 
-import android.os.Parcelable
-import androidx.compose.runtime.Immutable
-import kotlinx.android.parcel.Parcelize
+import androidx.compose.ui.graphics.Color
 
-@Immutable
-@Parcelize
-data class Poster(
-  val name: String,
-  val release: String,
-  val playtime: String,
-  val description: String,
-  val poster: String
-) : Parcelable
+val purple200 = Color(0xFF651FFF)
+val purple500 = Color(0xFF6200EA)
+val background = Color(0xFF2B292B)
+val background800 = Color(0xFF424242)
+val background900 = Color(0xFF212121)
+val white87 = Color(0Xddffffff)
