+++ app/src/main/java/com/github/skydoves/fisherdemo/ui/MainPosters.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.github.skydoves.frescomposabledemo.ui
+package com.github.skydoves.fisherdemo.ui
 
 import androidx.compose.foundation.ScrollableColumn
 import androidx.compose.foundation.Text
@@ -34,11 +34,11 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.ui.tooling.preview.Preview
-import com.github.skydoves.frescomposabledemo.model.MockUtil
-import com.github.skydoves.frescomposabledemo.model.Poster
-import com.github.skydoves.frescomposabledemo.theme.DisneyComposeTheme
-import com.github.skydoves.frescomposabledemo.theme.purple500
-import com.skydoves.frescomposable.FrescoImage
+import com.github.skydoves.fisherdemo.model.MockUtil
+import com.github.skydoves.fisherdemo.model.Poster
+import com.github.skydoves.fisherdemo.theme.DisneyComposeTheme
+import com.github.skydoves.fisherdemo.theme.purple500
+import com.skydoves.fisher.fresco.FrescoImage
 
 @Composable
 fun DisneyPosters(
