+++ app/src/main/res/values-night/themes.xml
@@ -1,6 +1,6 @@
 <resources xmlns:tools="http://schemas.android.com/tools">
   <!-- Base application theme. -->
-  <style name="Theme.FrescomposableDemo" parent="Theme.MaterialComponents.DayNight.DarkActionBar">
+  <style name="Theme.Fisher" parent="Theme.MaterialComponents.DayNight.DarkActionBar">
     <!-- Primary brand color. -->
     <item name="colorPrimary">@color/purple_200</item>
     <item name="colorPrimaryVariant">@color/purple_700</item>
