+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-  <string name="app_name">FrescomposableDemo</string>
+  <string name="app_name">Fisher</string>
 </resources>
