+++ app/src/main/res/values/themes.xml
@@ -1,6 +1,6 @@
 <resources xmlns:tools="http://schemas.android.com/tools">
   <!-- Base application theme. -->
-  <style name="Theme.FrescomposableDemo" parent="Theme.MaterialComponents.DayNight.DarkActionBar">
+  <style name="Theme.Fisher" parent="Theme.MaterialComponents.DayNight.DarkActionBar">
     <!-- Primary brand color. -->
     <item name="colorPrimary">@color/purple_500</item>
     <item name="colorPrimaryVariant">@color/purple_700</item>
@@ -14,7 +14,7 @@
     <!-- Customize your theme here. -->
   </style>
 
-  <style name="Theme.FrescomposableDemo.NoActionBar">
+  <style name="Theme.Fisher.NoActionBar">
     <item name="windowActionBar">false</item>
     <item name="windowNoTitle">true</item>
   </style>
