+++ settings.gradle
@@ -1,3 +1,3 @@
-rootProject.name = "FrescomposableDemo"
+rootProject.name = "FisherDemo"
 include ':app'
-include ':frescomposable'
+include ':fresco'
